Fishery Resilience 1 update
Clarified that this resilience is estimated at the permit level. Based off of EOP recommendations, also including scallop and clams in revenue as originally intended. This means I switched mfrow from c(2,1) to c(1,1), and commented out the second soe.plot commands. Also changed commercial revenue plot to indicate that this is in $2015.